backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx==0.28.1
langchain==0.3.14
langchain-openai==0.3.1
openai==1.59.9
requests==2.32.3